@@ -34,29 +34,31 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^5.2.9",
-        "@angular/common": "^5.2.9",
-        "@angular/compiler": "^5.2.9",
-        "@angular/core": "^5.2.9",
-        "@angular/forms": "^5.2.9",
-        "@angular/http": "^5.2.9",
-        "@angular/platform-browser": "^5.2.9",
-        "@angular/platform-browser-dynamic": "^5.2.9",
-        "@angular/platform-server": "^5.2.9",
-        "@angular/router": "^5.2.9",
+        "@angular/animations": "^6.1.9",
+        "@angular/common": "^6.1.9",
+        "@angular/compiler": "^6.1.9",
+        "@angular/core": "^6.1.9",
+        "@angular/forms": "^6.1.9",
+        "@angular/http": "^6.1.9",
+        "@angular/platform-browser": "^6.1.9",
+        "@angular/platform-browser-dynamic": "^6.1.9",
+        "@angular/platform-server": "^6.1.9",
+        "@angular/router": "^6.1.9",
         "bootstrap": "^3.3.7",
         "core-js": "^2.4.1",
         "jquery": "^3.2.1",
         "json2typescript": "^1.0.5",
         "ngx-bootstrap": "^2.0.3",
-        "rxjs": "^5.5.6",
+        "rxjs": "^6.3.3",
+        "rxjs-compat": "^6.0.0-rc.0",
         "snapsvg": "^0.5.1",
-        "zone.js": "^0.8.19"
+        "zone.js": "^0.8.26"
     },
     "devDependencies": {
-        "@angular/cli": "^1.7.3",
-        "@angular/compiler-cli": "^5.2.9",
-        "@angular/language-service": "^5.2.9",
+        "@angular-devkit/build-angular": "~0.8.0",
+        "@angular/cli": "^6.2.4",
+        "@angular/compiler-cli": "^6.1.9",
+        "@angular/language-service": "^6.1.9",
         "@types/jasmine": "~2.8.3",
         "@types/jasminewd2": "~2.0.2",
         "@types/node": "~6.0.60",
@@ -76,7 +78,7 @@
         "standard-version": "^4.4.0",
         "ts-node": "~4.1.0",
         "tslint": "~5.9.1",
-        "typescript": "~2.5.3",
-        "webpack-bundle-analyzer": "^3.0.1"
+        "typescript": "~2.9.2",
+        "webpack-bundle-analyzer": "^3.0.2"
     }
 }
