@@ -1,7 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { DataStreamerService } from '@awg-core/services';
 import { SearchResponseWithQuery } from '@awg-views/data-view/models';
