@@ -38,7 +38,7 @@ export class RouterStub {
 
 // Only implements params and part of snapshot.params
 // #docregion activated-route-stub
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { BehaviorSubject } from 'rxjs';
 
 @Injectable()
 export class ActivatedRouteStub {
